Time: 0 ms (100.00%), Space: 5.9 MB (92.49%) - LeetHub

+++ 1802-maximum-value-at-a-given-index-in-a-bounded-array/1802-maximum-value-at-a-given-index-in-a-bounded-array.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     long long getSumByFormula(long long count, long long val){
         
-        return (long long)(count * val) - (long long)(count * (count+1))/2;
+        return (count * val) - (count * (count+1))/2;
     }
     
     int better(int n, int index, int maxSum){
